Server/requirements.txt
alembic==1.7.6
asn1crypto==1.5.0
beautifulsoup4==4.10.0
blinker==1.4
boto3==1.21.18
botocore==1.24.18
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.1
cycler==0.11.0
Flask==2.0.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.1
importlib-resources==5.4.0
imutils==0.5.4
itsdangerous==2.1.0
Jinja2==3.0.3
jmespath==0.10.0
jsonify==0.5
lxml==4.8.0
Mako==1.1.6
MarkupSafe==2.1.0
mysql-connector-python==8.0.28
packaging==21.3
protobuf==3.19.4
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyngrok==5.1.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
redshift-connector==2.0.905
requests==2.27.1
s3transfer==0.5.2
scramp==1.4.1
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.31
typing_extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
zipp==3.7.0
